# modules/swagger-jaxrs-client-kek/swagger-inputs/swagger.yaml
swagger: "2.0"
info:
  description: "BaaS for delivery services like Glovo, Uber or even for regular mail. Also, a simple web front-end should be provided as an example of a typical consumer's app."
  version: "1.0.0"
  title: "Kinda Express King"
#  contact:
#    email: "apiteam@softserveinc.com"
# host: "localhost:8080"
...
paths:
...
  /orders:
    get:
      tags:
        - "orders"
      summary: "Searches for orders"
      description: "Finds all orders"
      operationId: "getOrderList"
      produces:
        - "application/vnd.softserve.orderList+json"
      responses:
        200:
          description: "The order object"
          schema:
            $ref: "#/definitions/orderList"
    post:
      tags:
        - "orders"
      summary: "Creates a new order"
      operationId: "addOrder"
      consumes:
        - "application/vnd.softserve.order+json"
      produces:
        - "application/vnd.softserve.order+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "order"
          in: "body"
          description: "The order to create"
          schema:
            $ref: "#/definitions/order"
      responses:
        200:
          description: "The newly created order object"
          schema:
            $ref: "#/definitions/order"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
  /orders/{guid}:
    get:
      tags:
        - "orders"
      summary: "Finds the specific order"
      operationId: "getOrder"
      produces:
        - "application/vnd.softserve.order+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific order"
          required: true
          type: "string"
      responses:
        200:
          description: "The order object"
          schema:
            $ref: "#/definitions/order"
    put:
      tags:
        - "orders"
      summary: "Modifies the specific order"
      operationId: "modifyOrder"
      consumes:
        - "application/vnd.softserve.order+json"
      produces:
        - "application/vnd.softserve.order+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific order"
          required: true
          type: "string"
        - name: "order"
          in: "body"
          description: "The order to modify"
          schema:
            $ref: "#/definitions/order"
      responses:
        200:
          description: "The modified order object"
          schema:
            $ref: "#/definitions/order"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
    delete:
      tags:
        - "orders"
      summary: "Deletes the specific order"
      operationId: "deleteOrder"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific order"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
  /orders/{guid}/events:
    get:
      tags:
        - "orders"
      summary: "Finds the specific order events"
      operationId: "getEvents"
      produces:
        - "application/vnd.softserve.eventList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific order"
          required: true
          type: "string"
      responses:
        200:
          description: "The list of order event objects"
          schema:
            $ref: "#/definitions/orderEventList"
    post:
      tags:
        - "orders"
      summary: "Adds an order event"
      operationId: "addEvent"
      consumes:
        - "application/vnd.softserve.event+json"
      produces:
        - "application/vnd.softserve.event+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific order"
          required: true
          type: "string"
        - name: "event"
          in: "body"
          description: "The order event to add"
          schema:
            $ref: "#/definitions/orderEvent"
      responses:
        200:
          description: "The list of order event objects"
          schema:
            $ref: "#/definitions/orderEvent"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
definitions:
...
  orderList:
    type: "object"
    properties:
      orderList:
        type: "array"
        items:
          $ref: "#/definitions/order"
  order:
    type: "object"
    properties:
      tenant:
        type: "string"
        description: "tenant id"
      user:
        type: "string"
        description: "user id"
      guid:
        type: "string"
      summary:
        type: "string"
      details:
        $ref: "#/definitions/orderDetails"
  orderDetails:
    type: "object"
    properties:
      payload:
        type: "string"
      imageURL:
        type: "string"
  orderEventList:
    type: "object"
    properties:
      orderEventList:
        type: "array"
        items:
          $ref: "#/definitions/orderEvent"
  orderEvent:
    type: "object"
    properties:
      guid:
        type: "string"
      order:
        type: "string"
      type:
        type: "string"
        description: "order event type"
        enum:
          - "created"
          - "assigned"
          - "started"
          - "delivered"
      payload:
        type: "string"
...
  errorList:
    type: "object"
    properties:
      errorlList:
        type: "array"
        items:
          $ref: "#/definitions/error"
  error:
    type: "object"
    properties:
      error:
        type: "string"
      details:
        type: "string"

# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /burnkr/kek/1.0.0
schemes:
  - https